test env setup
init babel, jest
