+++ app/package.json
@@ -9,6 +9,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "express-winston": "^4.2.0",
     "joi": "^17.6.0",
@@ -19,5 +20,12 @@
     "winston-daily-rotate-file": "^4.7.1",
     "winston-elasticsearch": "^0.17.1",
     "winston-mongodb": "^5.0.7"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.18.5",
+    "@babel/preset-env": "^7.18.2",
+    "babel-jest": "^28.1.1",
+    "superagent": "^7.1.6",
+    "supertest": "^6.2.3"
   }
 }
